Fix the neon glow, and share the tool popups with the paged editor
The halo was drawn at full device scale into a buffer half the size, so
it landed magnified and anchored at the corner rather than over the
stroke. The zoom now shrinks with the buffer, and since the buffer is
rounded up to whole pixels the composite samples only the part the
viewport actually occupied, so no fraction of a pixel slides the halo off
what it belongs to.

The stroke under the pen was invisible until it committed because the wet
path kept only the first of an item's runs and dropped its glow spec. A
neon stroke is three runs: a halo, a lit body and a white core. It
publishes all of them now, and the wet buffer is re-bound after a glow
pass rather than being left pointing at the committed one.

The tool popups are now the paged editor's own composables rather than
lookalikes, reached through a small interface both editors implement. A
lookalike drifts: a control added to one, a label reworded, a range
widened, and the two stop matching in ways nobody notices until a user
does. Identical is a property of the code now rather than something to
keep checking.

Pen styles, the shape style and the ink swatches are the same persisted
settings on both surfaces too, so a pen tuned on a canvas is that pen on
a note, and neon survives a relaunch.

+++ app/src/main/java/com/xnotes/gl/InkShader.kt
@@ -36,8 +36,8 @@ class InkShader(contextGen: Int) {
         localScrollX: Double,
         localScrollY: Double,
         zoom: Double,
-        viewportW: Int,
-        viewportH: Int,
+        viewportW: Double,
+        viewportH: Double,
     ) {
         program.use()
         program.set("uCamChunk", camChunkX.toFloat(), camChunkY.toFloat())
